Fix/Improvements

+++ app/Http/Controllers/TaskController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Contracts\PayloadService;
 use App\Http\Requests\Task\TaskRequest;
 use App\Models\Category;
 use App\Models\Task;
@@ -17,16 +18,18 @@ class TaskController extends Controller
     {
         $user  = auth()->user();
         $tasks = Task::whereBelongsTo($user)
-            ->orderBy('due_date')
-            ->get();
+            ->with(['category'])
+            ->orderBy('due_date');
 
         if ($request->input('status')) {
             $tasks = $tasks->where('status', $request->input('status'));
         }
 
-        if ($request->input('frequência')) {
-            $tasks = $tasks->where('frequency', $request->input('frequência'));
+        if ($request->input('frequency')) {
+            $tasks = $tasks->where('frequency', $request->input('frequency'));
         }
+
+        $tasks = $tasks->paginate(5);
 
         return view('dashboard', compact(['tasks']));
     }
@@ -36,30 +39,23 @@ class TaskController extends Controller
      */
     public function create()
     {
-        $categories = Category::where('user_id', Auth::id())->get();
-        if ($categories->isEmpty()) {
-            return redirect('dashboard')->with('failed', 'Nenhuma categoria encontrada');
-        } else {
-            return view('task.form', ['task' => new Task(), 'categories' => $categories]);
-        }
+        $user       = auth()->user();
+        $categories = Category::whereBelongsTo($user)->get();
+
+        abort_unless($categories->isNotEmpty(), 403, 'Categoria não encontrada');
+
+        return view('task.form', ['tasks' => new Task(), 'categories' => $categories]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(TaskRequest $request)
+    public function store(TaskRequest $request, Task $task)
     {
+
         $validated = $request->validated();
 
-        Task::create([
-            'title'       => $validated['title'],
-            'description' => $validated['description'],
-            'due_date'    => $validated['due_date'],
-            'status'      => $validated['status'],
-            'frequency'   => $validated['frequency'],
-            'user_id'     => Auth::id(),
-            'category_id' => $validated['category_id'],
-        ]);
+        app(PayloadService::class, ['payload' => $validated, 'model' => 'Task'])->create($task);
 
         return redirect('dashboard')->with('success', 'Tarefa criada com sucesso!');
     }
@@ -67,37 +63,33 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(string $id, Category $category)
     {
-        $task       = Task::where('id', $id)->where('user_id', Auth::id())->firstorFail();
-        $categories = Category::where('user_id', Auth::id())->get();
+        $task       = Task::findOrFail($id);
+        $user       = auth()->user();
 
-        return view('task.show', compact(['task', 'categories']));
-    }
+        abort_unless($user->can('update', $task), 403);
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
+        $categories = Category::whereBelongsTo($user)->get();
+
+        return view('task.show', compact('task', 'categories'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(TaskRequest $request, string $id)
     {
-        Task::where('id', $id)
-            ->update([
-                'title'       => $request->title,
-                'description' => $request->description,
-                'due_date'    => $request->due_date,
-                'status'      => $request->status,
-                'frequency'   => $request->frequency,
-            ]);
+        $user       = Auth::user();
+        $task       = Task::findOrFail($id);
 
-        return redirect('dashboard')->with('success', 'Tarefa atualizada');
+        abort_unless($user->can('update', $task), 403);
+
+        $validated = $request->validated();
+
+        app(PayloadService::class, ['payload' => $validated, 'model' => 'Task'])->update($task);
+
+        return redirect()->back()->with('success', 'Tarefa atualizada');
     }
 
     /**
@@ -105,8 +97,13 @@ class TaskController extends Controller
      */
     public function destroy(string $id)
     {
+        $user  = auth()->user();
+        $task  = Task::findOrFail($id);
+
+        abort_unless($user->can('destroy', $task), 403);
+
         Task::destroy($id);
 
-        return redirect('dashboard')->with('success', 'Tarefa deletada');
+        return redirect()->back()->with('success', 'Tarefa deletada');
     }
 }
